2024 Day 20 - Optimized Solution
I didn't check in the original solution that ran in about 6-7seconds
time. This solution runs in about 1.0 to 1.3 seconds on average.

Initial part 1 solution was very different and got me too stuck on
validating that cheats go through walls, or at most one wall. Going
through walls is necessary to cheat but it turns out that's more of side
effect than a requirement of what needs to be checked.