2024 Day 21
After solving a few other days, recovering from ilness, and studying
some threads on reddit, it clicked what I needed for a DFS solution. I
had tried a few things but the numbers were always too small. I was
missing my logic for "chunking" the input and inserting the extra
implicit A's.

This is a really interesting puzzle.